requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
executing==2.2.1
fastapi==0.116.1
fastjsonschema==2.21.2
Flask==3.1.2
flask-cors==6.0.1
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
idna==3.10
ipykernel==6.30.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
PyJWT==2.8.0
bcrypt==4.1.2
pytesseract==0.3.10
Pillow==10.1.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
motor==3.7.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.15.5
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.3
tinycss2==1.4.0
titlecase==2.4.1
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
yarg==0.1.9
